+++ aim/web/ui/src/services/models/imagesExplore/imagesExploreAppModel.ts
@@ -69,6 +69,7 @@ import { formatToPositiveNumber } from 'utils/formatToPositiveNumber';
 import getMinAndMaxBetweenArrays from 'utils/getMinAndMaxBetweenArrays';
 import getTooltipData from 'utils/app/getTooltipData';
 import filterTooltipContent from 'utils/filterTooltipContent';
+import { getDataAsMediaSetNestedObject } from 'utils/app/getDataAsMediaSetNestedObject';
 import { getCompatibleSelectConfig } from 'utils/app/getCompatibleSelectConfig';
 import { getValue } from 'utils/helper';
 
@@ -438,10 +439,11 @@ function setModelData(rawData: any[], configData: IImagesExploreAppConfig) {
       contexts,
     }),
   ];
-  const { imageSetData, orderedMap } = getDataAsImageSet(
+  const { setData, orderedMap } = getDataAsMediaSetNestedObject({
     data,
     groupingSelectOptions,
-  );
+    model,
+  });
 
   tooltipData = getTooltipData({
     processedData: data,
@@ -513,7 +515,7 @@ function setModelData(rawData: any[], configData: IImagesExploreAppConfig) {
     config,
     params,
     data,
-    imagesData: imageSetData,
+    imagesData: setData,
     orderedMap,
     tableData: tableData.rows,
     tableColumns: getImagesExploreTableColumns(
@@ -543,10 +545,11 @@ function updateModelData(
       contexts,
     }),
   ];
-  const { imageSetData, orderedMap } = getDataAsImageSet(
+  const { setData, orderedMap } = getDataAsMediaSetNestedObject({
     data,
     groupingSelectOptions,
-  );
+    model,
+  });
   tooltipData = getTooltipData({
     processedData: data,
     paramKeys: sortedParams,
@@ -600,7 +603,7 @@ function updateModelData(
   model.setState({
     config: configData,
     data: model.getState()?.data,
-    imagesData: imageSetData,
+    imagesData: setData,
     orderedMap,
     // chartTitleData: getChartTitleData(data),
     tableData: tableData.rows,