+++ aim/web/ui/src/services/models/runs/util.ts
@@ -1,5 +1,5 @@
 import bs58check from 'bs58check';
-import { head, orderBy } from 'lodash-es';
+import _ from 'lodash-es';
 
 import { IMenuItem } from 'components/kit/Menu';
 
@@ -12,6 +12,7 @@ import contextToString from 'utils/contextToString';
 import alphabeticalSortComparator from 'utils/alphabeticalSortComparator';
 import { encode } from 'utils/encoder/encoder';
 import getObjectPaths from 'utils/getObjectPaths';
+import { getDataAsMediaSetNestedObject } from 'utils/app/getDataAsMediaSetNestedObject';
 import { getValue } from 'utils/helper';
 
 import imagesExploreAppModel from '../imagesExplore/imagesExploreAppModel';
@@ -304,12 +305,56 @@ export function processImagesData(
       });
     });
   });
-  const { imageSetData, orderedMap } = imagesExploreAppModel.getDataAsImageSet(
-    groupData(orderBy(images)),
+  const { setData, orderedMap } = getDataAsMediaSetNestedObject({
+    data: groupData(_.orderBy(images)),
     groupingSelectOptions,
-    ['step'],
-  );
-  return { imageSetData, orderedMap, record_range, index_range };
+    defaultGroupFields: ['step'],
+  });
+  return { imageSetData: setData, orderedMap, record_range, index_range };
+}
+
+export function processAudiosData(
+  data: Partial<ImagesData>,
+  params?: { [key: string]: unknown },
+) {
+  const { record_range, iters, values, index_range, context, name } = data;
+  const groupingSelectOptions = params
+    ? imagesExploreAppModel.getGroupingSelectOptions({
+        params: getObjectPaths(params, params),
+      })
+    : [];
+  let audiosSetData: any[] = [];
+
+  values?.forEach((stepData: IImageData[], stepIndex: number) => {
+    stepData.forEach((audio: IImageData) => {
+      const audioKey = encode({
+        name,
+        traceContext: context,
+        index: audio.index,
+        step: iters?.[stepIndex],
+        caption: audio.caption,
+      });
+      const seqKey = encode({
+        name,
+        traceContext: context,
+      });
+      audiosSetData.push({
+        ...audio,
+        audio_name: name,
+        step: iters?.[stepIndex],
+        context: context,
+        key: audioKey,
+        seqKey: seqKey,
+      });
+    });
+  });
+  const { setData, orderedMap } = getDataAsMediaSetNestedObject({
+    data: groupData(_.orderBy(audiosSetData)),
+    groupingSelectOptions,
+    defaultGroupFields: ['step'],
+  });
+
+  return { audiosSetData: setData, orderedMap, record_range, index_range };
 }
 
 function groupData(data: IProcessedImageData[]): {
@@ -361,7 +406,7 @@ export function reformatArrayQueries(
  */
 export function processPlotlyData(data: Partial<IPlotlyData>) {
   const { record_range, iters, values } = data;
-  const processedValue = head(values);
+  const processedValue = _.head(values);
   const originalValues = values;
 
   processedValue.layout.autosize = true;